+++ lib/extended_math.dart
@@ -1,4 +1,5 @@
-/// Library that add functionality of all maths sections that don't exist in `dart:math`
+/// Library that add functionality of all maths sections that don't
+/// exist in `dart:math`
 ///
 /// At the moment library have 4 sections:
 ///
@@ -8,16 +9,19 @@
 ///     4. Applied mathematics
 ///
 /// Each section don't have full implementation yet.
-/// See [dartdoc](https://pub.dartlang.org/documentation/extended_math/latest/) for which functionality are implemented.
+/// See [dartdoc](https://pub.dartlang.org/documentation/extended_math/latest/)
+/// for which functionality are implemented.
 ///
-/// Sections are created according to [Mathematics Subject Classification](https://en.wikipedia.org/wiki/Mathematics_Subject_Classification).
+/// Sections are created according to
+/// [Mathematics Subject Classification](https://en.wikipedia.org/wiki/Mathematics_Subject_Classification).
 library extended_math;
 
+export 'src/applied_mathematics/probability_theory/continuous_probability_distributions/uniform_distribution.dart';
 export 'src/applied_mathematics/probability_theory/numbers_generator.dart';
-export 'src/applied_mathematics/probability_theory/probability_distributions/continuous_uniform_distribution.dart';
-export 'src/applied_mathematics/probability_theory/theory_of_probability_distributions.dart';
 export 'src/applied_mathematics/statistic/central_tendency.dart';
+export 'src/applied_mathematics/statistic/dispersion.dart';
 export 'src/applied_mathematics/statistic/exceptions/mean_exception.dart';
+export 'src/applied_mathematics/statistic/shape_of_probability_distribution.dart';
 
 export 'src/complex_analysis/complex.dart';